Animation with Angular
Animation with Angular
————————

Start by installing angular animations library with npm so we can
import it. We are using it as a dev dependency with the command ‘npm
install @angular/animations@latest —save’

Import animations to the component we want to animate. In the Home
component, we import a number of animation specific functions.

Define animations within the component code as property called
animations. Be sure to use ‘optional: true’ so that sometimes
animations will not fail.

Add animation to template by adding [@goals]="goals.length" into the
element in the HTML template. ‘goals’ is the name of the trigger that
we defined in the Home component.