Using the router
Using the router
—————————————

app-routing.module.ts file is what is set up when we use the —routing
flag when we generated the application. We use that file for working
with routing.

Start by importing the routes: import { HomeComponent } from
'./home/home.component';
import { AboutComponent } from './about/about.component';

Set up routes in Routes array in this file like this: const routes:
Routes = [
  {
    path: '',
    component: HomeComponent
  },
  {
    path: 'about',
    component: AboutComponent
  },
];

Make sure app.component.html has no routes listed in it to allow for
the clicking of each route to do the routing and it works!