Set and fetch route parameters in Angular5
Set and fetch route parameters in Angular5
——————————————

While this does not require query parameters, we’re going to create
them in order to understand them.

path: 'about/:id' is how it works and path might correspond to a
database entry or particular file path (starting with it hard coded)

Now we want to retrieve the route parameter in the application.

import ActivatedRoute from the angular library into the
about.component.ts file: import { ActivatedRoute } from
'@angular/router’;, which gives us access to the route parameters.

We use dependency injection to create an instance of the activated
route in the constructor within the about.component.ts file.

The constructor is called when the route is loaded.

this.route.params.subscribe(res => console.log(res.id)); shows us the
route id, what we defined as the route parameter in the
app-routing.module.ts file.

Typically, one would define something that deals with the parameter
above the constructor function in order to manipulate and/or use the
route parameter(s).