Using Angular5 Services
Using Angular5 Services
——————————————

How to create a service file to use data across multiple components.
These files are generally used to make HTTP calls and/or for sharing
data between components.

Generate a service with ng called ‘data’: ng g s <name>

In the service files, instead of a component with a component injector,
there is an injectable within the service file.

One of the best ways to share data between components is to use the
RSJX subject library.

providers: [DataService] gets added to the providers array in
app.module.ts file.

import DataService into home.component.ts file as well and then create
an instance of the dependency injection, like this: constructor(private
_data: DataService) { }

Now, in ngInit(), we have access to this._data.goal.subscribe(res =>
this.goals = res)

Now the ‘goals’ are being pulled from the service file, rather than the
component. This enables the goals to be shared between components.

Point of a service file(s) is that you can create it/them and then
import it/them into other components so that those components have
access to the properties and the methods of the service file(s).