Angular5 deployment
Angular5 deployment
—————————————

First step to take the application live is to create a build with the Angular5 CLI with ‘ng build’. Gives a 3.43 MB file set.

The command creates a dist folder. It’s big, so you don’t want to run it without the production flag. So we run ‘ng build —prod’. Creates a 796 byte file because it runs compression and optimization applications.

With the dist(ribution) folder created, we can now deploy it.

If deploying to a site on which the dist is a subfolder, you should run the command ‘bg build —prod —base-href=“your_url_here”

Can also use github pages. This can be done through a CLI tool. ‘npm install -g angular-cli-ghpages’ GitHub pages is actually a sub-folder, so we need to re-run our build command.

ng build --prod base-href="https://pokeredge.g
ithub.io/Angular-Bucket-List-Application/"

Finally, run the command ‘angular-cli-ghpages’